LPS-112856 search-experiences-web: Remove internal.blueprint package
